Added flag "Important" to chat message (raise the receiver chat window on top)
git-svn-id: https://svn.code.sf.net/p/beebeep/code@1064 127b15f6-8070-407b-945c-913c06b37495

+++ src/desktop/Main.cpp
@@ -114,7 +114,7 @@ int main( int argc, char *argv[] )
   Settings::instance().load();
   Settings::instance().createApplicationUuid();
   Log::instance().setMaxLogLines( Settings::instance().maxLogLines() );
-  qDebug() << "Log has set max lines to:" << (int)Settings::instance().maxLogLines();
+  qDebug() << "Log has set max lines to:" << Settings::instance().maxLogLines();
 #ifdef BEEBEEP_DEBUG
   QFont chat_font = Settings::instance().chatFont();
   qDebug() << "Font selected for chat:" << chat_font.toString();
